Use identity instead of auto increment
- rename auto_increment_test to identity_test
- use identity keyword instead of auto_increment keyword
